+++ src/components/molecules/AdressForm/AdressForm.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+import { Input } from 'components/atoms/Input/Input';
+
+export const TimeInput = styled(Input)`
+  display: flex;
+  flex-direction: row;
+`;
